refactor cubes

+++ src/dimension.rs
@@ -1,4 +1,4 @@
 pub struct Dimension {
-    pub width: i32,
-    pub height: i32,
+    pub width: usize,
+    pub height: usize,
 }
